Always do swiftmodule, somtimes also objects
Summary: apple_binary still uses swift-companion, which reaches this where module only is set to false. However, we have some binaries that use `UIApplicationMain` instead of having a main file. For now just always emit the module, as you need the module to write the object files anyway. Going to see if this fixes it.

Reviewers: vkal, richardh

Differential Revision: https://code.uberinternal.com/D1416235